Split out DrawTextWin32 into its own file, to be further split later.
