+++ ext/native/gfx_es2/draw_text.h
@@ -15,6 +15,7 @@
 
 #include "base/basictypes.h"
 #include "gfx_es2/draw_buffer.h"
+#include "util/text/wrap_text.h"
 
 namespace Draw {
 	class DrawContext;
@@ -46,11 +47,6 @@ enum {
 	FONTSTYLE_ITALIC = 2,
 };
 
-// Internal struct but all details in .cpp file (pimpl to avoid pulling in excessive headers here)
-struct TextDrawerContext;
-class TextDrawerFontContext;
-
-
 class TextDrawer {
 public:
 	TextDrawer(Draw::DrawContext *draw);
@@ -58,25 +54,39 @@ public:
 
 	virtual uint32_t SetFont(const char *fontName, int size, int flags) = 0;
 	virtual void SetFont(uint32_t fontHandle) = 0;  // Shortcut once you've set the font once.
-	virtual void SetFontScale(float xscale, float yscale) = 0;
+	void SetFontScale(float xscale, float yscale);
 	virtual void MeasureString(const char *str, size_t len, float *w, float *h) = 0;
 	virtual void MeasureStringRect(const char *str, size_t len, const Bounds &bounds, float *w, float *h, int align = ALIGN_TOPLEFT) = 0;
 	virtual void DrawString(DrawBuffer &target, const char *str, float x, float y, uint32_t color, int align = ALIGN_TOPLEFT) = 0;
 	virtual void DrawStringRect(DrawBuffer &target, const char *str, const Bounds &bounds, uint32_t color, int align) = 0;
 	// Use for housekeeping like throwing out old strings.
 	virtual void OncePerFrame() = 0;
-	virtual float CalculateDPIScale() = 0;
+
+	float CalculateDPIScale();
 
 protected:
 	Draw::DrawContext *draw_;
 	virtual void ClearCache() = 0;
 	virtual void RecreateFonts() = 0;  // On DPI change
-	virtual void WrapString(std::string &out, const char *str, float maxWidth) = 0;
+	void WrapString(std::string &out, const char *str, float maxWidth);
 
 	int frameCount_;
 	float fontScaleX_;
 	float fontScaleY_;
 	float dpiScale_;
 };
+
+
+class TextDrawerWordWrapper : public WordWrapper {
+public:
+	TextDrawerWordWrapper(TextDrawer *drawer, const char *str, float maxW) : WordWrapper(str, maxW), drawer_(drawer) {
+	}
+
+protected:
+	float MeasureWidth(const char *str, size_t bytes) override;
+
+	TextDrawer *drawer_;
+};
+
 
 TextDrawer *CreateTextDrawer(Draw::DrawContext *draw);
