+++ ext/native/gfx_es2/draw_text_win.h
@@ -1,0 +1,44 @@
+#pragma once
+
+#include "ppsspp_config.h"
+
+#include <map>
+#include "gfx_es2/draw_text.h"
+
+#if defined(_WIN32) && !defined(USING_QT_UI) && !PPSSPP_PLATFORM(UWP)
+
+#define WIN32_LEAN_AND_MEAN
+#include <Windows.h>
+
+class TextDrawerContext;
+// Internal struct but all details in .cpp file (pimpl to avoid pulling in excessive headers here)
+class TextDrawerFontContext;
+
+class TextDrawerWin32 : public TextDrawer {
+public:
+	TextDrawerWin32(Draw::DrawContext *draw);
+	~TextDrawerWin32();
+
+	uint32_t SetFont(const char *fontName, int size, int flags) override;
+	void SetFont(uint32_t fontHandle) override;  // Shortcut once you've set the font once.
+	void MeasureString(const char *str, size_t len, float *w, float *h) override;
+	void MeasureStringRect(const char *str, size_t len, const Bounds &bounds, float *w, float *h, int align = ALIGN_TOPLEFT) override;
+	void DrawString(DrawBuffer &target, const char *str, float x, float y, uint32_t color, int align = ALIGN_TOPLEFT) override;
+	void DrawStringRect(DrawBuffer &target, const char *str, const Bounds &bounds, uint32_t color, int align) override;
+	// Use for housekeeping like throwing out old strings.
+	void OncePerFrame() override;
+
+protected:
+	void ClearCache() override;
+	void RecreateFonts() override;  // On DPI change
+
+	TextDrawerContext *ctx_;
+	std::map<uint32_t, std::unique_ptr<TextDrawerFontContext>> fontMap_;
+
+	uint32_t fontHash_;
+	// The key is the CityHash of the string xor the fontHash_.
+	std::map<uint32_t, std::unique_ptr<TextStringEntry>> cache_;
+	std::map<uint32_t, std::unique_ptr<TextMeasureEntry>> sizeCache_;
+};
+
+#endif
