+++ ext/native/native.vcxproj.filters
@@ -314,6 +314,12 @@
     <ClInclude Include="thin3d\d3d11_loader.h">
       <Filter>thin3d</Filter>
     </ClInclude>
+    <ClInclude Include="gfx_es2\draw_text_win.h">
+      <Filter>gfx</Filter>
+    </ClInclude>
+    <ClInclude Include="gfx_es2\draw_text_qt.h">
+      <Filter>gfx</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="gfx\gl_debug_log.cpp">
@@ -765,6 +771,12 @@
     </ClCompile>
     <ClCompile Include="base\logging.cpp">
       <Filter>base</Filter>
+    </ClCompile>
+    <ClCompile Include="gfx_es2\draw_text_win.cpp">
+      <Filter>gfx</Filter>
+    </ClCompile>
+    <ClCompile Include="gfx_es2\draw_text_qt.cpp">
+      <Filter>gfx</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
